requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0

# WebSocket support
python-socketio==5.8.0
python-engineio==4.7.1
eventlet==0.33.3
deepgram-sdk==4.8.1

# HTTP and async support
requests==2.31.0
aiohttp==3.8.6

# Salesforce integration
simple-salesforce==1.12.4

# Slack integration
slack-sdk==3.21.3

# Security and authentication
PyJWT==2.8.0
cryptography==41.0.4

# Environment and configuration
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0

# Logging and monitoring
structlog==23.1.0

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1

# Data processing
pandas==2.0.3
pydantic==2.1.1

# Database (for future use)
psycopg2-binary==2.9.7
